% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_grid.R
\name{add_grid}
\alias{add_grid}
\title{CPL \link{ggplot2} theme}
\usage{
add_grid(flip = FALSE)
}
\arguments{
\item{flip}{TRUE/FALSE is \code{coord_flip()} used}
}
\description{
Adds major hoizontal panel gridlines
}
