% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_required_packages.R
\name{check_required_packages}
\alias{check_required_packages}
\title{A function that wraps up several lines of code to check for required packages.
If the required packages are not installed, then it will install them. This
requires automatically identifying the R version number (major, minor) and
combining that with the repo URL to use the correct url.}
\usage{
check_required_packages()
}
\value{

}
\description{
A function that wraps up several lines of code to check for required packages.
If the required packages are not installed, then it will install them. This
requires automatically identifying the R version number (major, minor) and
combining that with the repo URL to use the correct url.
}
