% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpl_make_tbl.R
\name{cpl_make_tbl}
\alias{cpl_make_tbl}
\title{CPL table styling with \link{gt}}
\usage{
cpl_make_tbl(data, tbl_num = NA, tbl_title = NA)
}
\arguments{
\item{data}{The table data for your table. \link{data.frame}}

\item{tbl_num}{The table number. \link{str}}

\item{tbl_title}{A descriptive table title. \link{str}}
}
\description{
Make a table conforming to CPL's table style
}
