% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpl_save.R
\name{cpl_save}
\alias{cpl_save}
\title{CPL \link{ggplot2} theme}
\usage{
cpl_save(filename, ...)
}
\arguments{
\item{filename}{The filename for your plot. \link{str}}
}
\description{
Save a ggplot with CPL defaults
}
