% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpl_set_theme.R
\name{cpl_set_theme}
\alias{cpl_set_theme}
\title{Makes the custom CPL theme the default theme for ggplot}
\usage{
cpl_set_theme(
  color_schema = "brief",
  base_size = 12,
  base_family = "Arial",
  base_line_size = 0.5,
  base_rect_size = 0.5
)
}
\arguments{
\item{base_size}{Base font size. Default is 12.}

\item{base_family}{Font. Default is Arial.}

\item{base_line_size}{Default is 0.5}

\item{base_rect_size}{Default is 0.5}
}
\value{

}
\description{
Makes the custom CPL theme the default theme for ggplot
}
