% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themeCPL.R
\name{customCPLTheme}
\alias{customCPLTheme}
\title{Constructs the custom CPL theme}
\usage{
customCPLTheme(
  base_size = 12,
  base_family = "Arial",
  base_line_size = 0.5,
  base_rect_size = 0.5
)
}
\arguments{
\item{base_size}{Base font size. Default is 12.}

\item{base_family}{Font. Default is Arial.}

\item{base_line_size}{Default is 0.5}

\item{base_rect_size}{Default is 0.5}
}
\value{
A theme that can then be used with \code{\link{theme_set}}
}
\description{
Constructs the custom CPL theme
}
