% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{func_cpl_palette}
\alias{func_cpl_palette}
\title{CPL [ggplot2] theme}
\usage{
func_cpl_palette(palette = COLOR_SCHEMA)
}
\arguments{
\item{palette}{Palette name.}
}
\description{
Applying CPL's color palettes as [ggplot2] defaults
}
