% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Test function to ensure the package has loaded}
\usage{
hello(name)
}
\arguments{
\item{name}{String to be concatenated into output}
}
\value{
The output from \code{\link{print}}
}
\description{
Test function to ensure the package has loaded
}
\examples{
hello('Landon')
}
