% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_axis.R
\name{remove_axis}
\alias{remove_axis}
\title{CPL \link{ggplot2} theme}
\usage{
remove_axis(axis = "y", flip = FALSE)
}
\arguments{
\item{axis}{Axis to drop. "x", "y", or "both"}

\item{flip}{TRUE/FALSE is \code{coord_flip()} used}
}
\description{
Removes axes and axes labels
}
