% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_discrete}
\alias{scale_fill_discrete}
\title{Discrete color scale that aligns with CPL style}
\usage{
scale_fill_discrete(...)
}
\arguments{
\item{...}{other arguments passed to \code{discrete_scale()}}
}
\description{
Discrete color scale that aligns with CPL style
}
