% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_axis.R
\name{remove_axis}
\alias{remove_axis}
\title{CPL \link{ggplot2} theme}
\usage{
remove_axis(axis = "y", flip = FALSE)
}
\arguments{
\item{axis}{Axis to drop. \link{str}: "x", "y", or "both"}

\item{flip}{Is \code{\link[=coord_flip]{coord_flip()}} used? \link{logical}: TRUE/FALSE}
}
\description{
Removes axes and axes' labels
}
